first commit
First commit, currently some shell scripts that check for memory usage  (lowmemallert.sh) and warn if it's low and reschedule ;processes if they start hogging IO or CPU. (maultipasszipo.sh). run with a nice of -3 so it's nice than the nice 0 processes sit monitors and changes. currently returns processes back to nice 0 if they stop hogging CPU but doesn't return IO scheduling after a processes hogged the IO.
This also needs to be turned into a service not a script
needs iotop installed, tested with Linux mint